Fix cgo-output of type-smallmoney
- Negative numbers of ASE-type `smallmoney` are misrepresented:
  Inserted value:     -999
  Value displayed:    428497.7296
  Explanation:        429496.7295 (maxInt32 without ".") - 999 = 428497.7296

  Thus, the byteslices are misinterpreted, which can be fixed by
  this adjustment.

+++ cgo/rows.go
@@ -227,8 +227,7 @@ func (rows *Rows) Next(dest []driver.Value) error {
 			if err != nil {
 				return fmt.Errorf("Received invalid precision/scale values from ASE: %v", err)
 			}
-
-			dec.SetInt64(int64(binary.LittleEndian.Uint32(bs)))
+			dec.SetInt64(int64(int32(binary.LittleEndian.Uint32(bs))))
 
 			dest[i] = dec
 		case types.DATE:
